test: split performance tests and normal test
